Extract listener for queryParam change into separate service (mainly useful for updates after browser back button)
